Use interval to create to scroll
FIREFOX ISSUE
According to the [spec](https://html.spec.whatwg.org/multipage/interaction.html#drag-and-drop-processing-model): User agents must, as soon as the drag operation is initiated and every 350ms (±200ms) thereafter for as long as the drag operation is ongoing, queue a task to perform the following steps in sequence. Firefox neglects firing events if dragged-object is over the drop element, but mouse is not moved.

Using a setInterval fixes the problem. Interval is started as soon dragged-object enters the drop element, interval is stopped if drop element is leaved.

BONUS FOR ALL BROWSERS
* The scroll is much more fluid
* The scroll does not go faster/slower if dragged-object is dragged fast over the drop element